changed a lot of backend names for consistency
