+++ limiter.go
@@ -1,0 +1,62 @@
+package main
+
+import (
+	"net/http"
+	"strings"
+	"sync"
+
+	"github.com/tedyst/spotifyutils/api/utils"
+	"golang.org/x/time/rate"
+)
+
+// Create a map to hold the rate limiters for each visitor and a mutex.
+var visitors = make(map[string]*rate.Limiter)
+var mu sync.Mutex
+
+// Retrieve and return the rate limiter for the current visitor if it
+// already exists. Otherwise create a new rate limiter and add it to
+// the visitors map, using the IP address as the key.
+func getVisitor(username string) *rate.Limiter {
+	mu.Lock()
+	defer mu.Unlock()
+
+	limiter, exists := visitors[username]
+	if !exists {
+		// Get a maximum of 8 api requests, getting a new one every 3 seconds
+		limiter = rate.NewLimiter(0.3, 8)
+		visitors[username] = limiter
+	}
+
+	return limiter
+}
+
+type limiterError struct {
+	Success bool
+	Error   string
+}
+
+func limitAPIRequests(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		// Get the IP address for the current user.
+		ipAddress := r.RemoteAddr
+		fwdAddress := r.Header.Get("X-Forwarded-For")
+		if fwdAddress != "" {
+			ipAddress = fwdAddress
+			ips := strings.Split(fwdAddress, ", ")
+			if len(ips) > 1 {
+				ipAddress = ips[0]
+			}
+		}
+
+		// Call the getVisitor function to retreive the rate limiter for
+		// the current user.
+		limiter := getVisitor(ipAddress)
+		if !limiter.Allow() {
+			utils.ErrorString(w, r, "Too many requests")
+			w.WriteHeader(http.StatusTooManyRequests)
+			return
+		}
+
+		next.ServeHTTP(w, r)
+	})
+}
